Mid Version Update
YOU MUST REBULD ANY SIMULATIONS CREATED WITH VERSIONS BEFORE V1.30.  This is due to changes in the simulation tab structure.

In this version, I have simplified the Parameters Table significantly. The run state and histograms values have been set to a new default that would have been the old values of "Fast" and "This Book," respectively.

The Date Time Stamp is now included in the name of any histograms. This helps when copying histograms to PPT because they can be compared for each run.

I also moved all deprecated code to a separate module for posterity's sake.

PREVIOUS CHANGES:
Minor function increase: Added line to display the formula as plain text below the equation line. This is to facilitate auditing and debugging.

The "About" box shows version number 1.30.1

PREVIOUS CHANGES TO VERSION 1.30
Fixed 2 bugs:
1. The Clean Formula function was inefficient and missed references that began with "http". It is now more efficient and catches a wider range of errors.
2. The new method for writing a run table by pasting the array previously overwrote the "Run Table:" label.  This is now fixed.

This release includes functional improvements from the base 1.30 release:
1. the ability to make the function in cell n dependent on a selectable value from a previous cell (n-m). You can use functions like Max, Min, Index, etc on the results of a previous function by selecting the range from row 25 to row 1048576 in the column of the previous function as the range or array. In order to do this, you must set Dependent  Trials to Y in the parameters table.
2. It is possible to have a function n dependent on the previous value of a later function. See the Example tab for instructions.
3. Updated the About dialog box
4. Converted the inc(n) function from Integer to Long. Remember that you now have 10 incrementing function that are called by enclosing the number of the function you want to use in the parentheses. For example "=inc(3)" will increment the third version of increment each time it is calculated.  You also have 10 functions that alternate between 1 and 0 each time they are recalculated.  They are called by "=bool(n)"
5. Cleaned up the code modules.
6. The date-time stamp of the last run of the simulation is written to the line below the parameters table every time you run the simulation.  This helps the tool meet the SIP standard as proposed by probabilitymanagement.org